[ANY] Rename freyr to freyja
