+++ packages/colors/lib/carbon/standard-base.css
@@ -1,8 +1,8 @@
 :root,
 .ui-standard {
-    --primary: #9747FF;
+    --primary: #222326;
 
-    --secondary: #232427;
+    --secondary: #484A4B;
 
     --danger: #f5385a;
