+++ packages/freyja/bin/freyja.js
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+import("../dist/freyja.js");
