+++ packages/freyja/eslint.config.js
