+++ packages/freyja/lib/cli.ts
