+++ packages/freyja/lib/constants.ts
@@ -4,7 +4,7 @@ const {version} = JSON.parse(
   readFileSync(new URL("../package.json", import.meta.url)).toString(),
 );
 
-export const FREYR_VERSION = version as string;
+export const FREYJA_VERSION = version as string;
 
 export const AUTO_GENERATE_DISCLAIMER = [
   "/*",
@@ -16,10 +16,10 @@ export const AUTO_GENERATE_DISCLAIMER = [
 export const BUTTON_SHADE_NAMES = ["-minor-2", "-minor-1", "", "-major-1", "-major-2", "-major-3", "-contrast"];
 
 export const CONFIG_FILES = [
-  "freyr.config.js",
-  "freyr.config.mjs",
-  "freyr.config.cjs",
-  "freyr.config.ts",
-  "freyr.config.mts",
-  "freyr.config.cts"
+  "freyja.config.js",
+  "freyja.config.mjs",
+  "freyja.config.cjs",
+  "freyja.config.ts",
+  "freyja.config.mts",
+  "freyja.config.cts"
 ];
