+++ packages/freyja/lib/logger/helpers.ts
