--- packages/freyr/bin/freyr.js
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-
-import("../dist/freyr.js");
